feat: add support for disabling events
This allows us to remove events from functions in the same way that we
can currently remove functions, with the very notable change that the
keyword that we look for is 'disabled' instead of 'enabled'.

We do this because in the AWS provider, the 'enabled' keyword already
exists for some event types, with an existing meaning.